pyproject.toml
[project]
name = "pokedex-scraper"
version = "0.1.0"
description = ""
authors = [
    {name = "WoutVossen",email = "vossen.w@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "scrapy (>=2.13.0,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "markdownify (>=1.1.0,<2.0.0)",
    "google-genai (>=1.15.0,<2.0.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymilvus (>=2.5.8,<3.0.0)"
]

[tool.poetry]
packages = [{include = "pokedex_scraper", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
